requirements/main.txt
# In a real project dev dependencies should be separated
Django==3.2.3
factory-boy==2.12.0
pytest==6.2.4
pytest-django==4.3.0
pytest-env==0.6.2
pytest-responses==0.5.0
psycopg2==2.8.6
djangorestframework==3.12.4
uWSGI==2.0.18
pytest-asyncio
pytest-deadfixtures
pytest-django
pytest-env
pytest-freezegun
pytest-mock
pytest-randomly
pytest-responses
pytest-xdist
